Education and EducationView converted to functional components
Tested and functional :smile:
