+++ services/frontend-v3/src/components/education/EducationView.jsx
@@ -1,13 +1,22 @@
 import React, { Component } from "react";
 import { injectIntl } from "react-intl";
-import moment from "moment";
 
 import { Icon as LegacyIcon } from "@ant-design/compatible";
 
 import { Row, Col, Avatar, List } from "antd";
 
-class EducationView extends Component {
-  generateEducationInfoList(dataSource) {
+function EducationView(props) {
+  /* Component Styles */
+  const styles = {
+    card: {
+      height: "100%"
+    },
+    avatar: {
+      backgroundColor: "#007471"
+    }
+  };
+
+  const generateEducationInfoList = dataSource => {
     return (
       <List
         itemLayout="horizontal"
@@ -30,77 +39,17 @@ class EducationView extends Component {
         )}
       />
     );
-  }
+  };
 
-  getEducationInfo(locale) {
-    const data = this.props.data;
+  const educationInfo = props.educationInfo;
 
-    let educationInfo = [];
-    if (data.education != null) {
-      data.education.forEach(educElement => {
-        const startDate = educElement.startDate[locale];
-        const endDate = educElement.endDate[locale];
-
-        const education = {
-          icon: "bank",
-          diploma: educElement.diploma.description[locale],
-          school: educElement.school.description[locale],
-          duration: this.getEducationDuration(startDate, endDate)
-        };
-
-        educationInfo.push(education);
-      });
-    }
-
-    return [...educationInfo];
-  }
-
-  getEducationDuration(startDate, endDate) {
-    const formatedStartDate = moment(startDate).format("ll");
-    const formatedEndDate = moment(endDate).format("ll");
-
-    const dateNotProvided = this.props.intl.formatMessage({
-      id: "profile.date.not.provided"
-    });
-    const present = this.props.intl.formatMessage({
-      id: "profile.end.date.present"
-    });
-
-    let duration = "";
-
-    if (startDate === null && endDate === null) {
-      duration = duration + dateNotProvided;
-    } else if (startDate !== null && endDate === null) {
-      duration = duration + formatedStartDate + " - " + present;
-    } else {
-      duration = duration + formatedStartDate + " - " + formatedEndDate;
-    }
-
-    return duration;
-  }
-  render() {
-    const locale = this.props.intl.formatMessage({ id: "language.code" });
-
-    const educationInfo = this.getEducationInfo(locale);
-
-    return (
-      <Row>
-        <Col xs={24} lg={24}>
-          {this.generateEducationInfoList(educationInfo)}
-        </Col>
-      </Row>
-    );
-  }
+  return (
+    <Row>
+      <Col xs={24} lg={24}>
+        {generateEducationInfoList(educationInfo)}
+      </Col>
+    </Row>
+  );
 }
-
-/* Component Styles */
-const styles = {
-  card: {
-    height: "100%"
-  },
-  avatar: {
-    backgroundColor: "#007471"
-  }
-};
 
 export default injectIntl(EducationView);
